- Fixed bug where IGNs were only world-specific  - Corrected min password size  - Fixed bug where experience gains would error 38 the client  - Overhauled the database schema; made everything consistent  - Removed the storage_items table, I feel that having basically the same schema in two places was a bad idea  - Fixed pet IDs to be 64-bit instead of 32-bit  - Zeroed out the SQL files because I can
git-svn-id: svn://70.113.44.214/VanaMirror/branches/buis_vana_overhaul@3086 09c1253f-3d09-8e40-9719-4320e56601b5

+++ src/ChannelServer/Pet.cpp
@@ -39,7 +39,7 @@ Pet::Pet(Player *player, Item *item) :
 	mysqlpp::Query query = Database::getCharDB().query();
 	query << "INSERT INTO pets (name) VALUES (" << mysqlpp::quote << this->name << ")";
 	mysqlpp::SimpleResult res = query.execute();
-	this->id = (int32_t) res.insert_id();
+	this->id = (int64_t) res.insert_id();
 	item->setPetId(this->id);
 }
